remove patches, consolidate gemfiles
