--- stash/stash_datacite/Gemfile
@@ -1,28 +1,0 @@
-source 'https://rubygems.org'
-
-gemspec
-gem 'parallel_tests', group: %i[development test]
-
-path '..' do
-  # gem 'stash_discovery'
-  gem 'stash_engine'
-  gem 'stash-sword'
-  gem 'stash-wrapper'
-end
-
-# group :development, :local_dev, :test do
-#   gem 'byebug'
-#   gem 'colorize'
-#   gem 'database_cleaner'
-#   gem 'diffy'
-#   gem 'equivalent-xml'
-#   gem 'factory_bot_rails', require: false
-#   gem 'nokogiri'
-#   gem 'rspec', '~> 3.5'
-#   gem 'rspec-rails', '~> 3.5'
-#   gem 'rubocop', '0.85.1'
-#   gem 'scss_lint'
-#   gem 'simplecov'
-#   gem 'simplecov-console'
-#   gem 'webmock'
-# end
