--- stash/stash_datacite/lib/stash_datacite/engine.rb
@@ -1,10 +1,0 @@
-require 'mysql2'
-require 'responders'
-require 'leaflet-rails'
-require 'datacite/mapping'
-require 'kaminari'
-module StashDatacite
-  class Engine < ::Rails::Engine
-    isolate_namespace StashDatacite
-  end
-end
