--- stash/stash_datacite/lib/stash_datacite/user_patch.rb
@@ -1,13 +1,0 @@
-module StashDatacite
-  # Ensure Author-Affiliation relation & related methods
-  # invoked from StashDatacite::Resource::Completions when first needed
-  module UserPatch
-
-    def self.patch!
-      StashEngine::User.instance_eval do
-        belongs_to :affiliation, class_name: 'StashDatacite::Affiliation', optional: true
-      end
-    end
-
-  end
-end
