--- stash/stash_datacite/lib/stash_datacite/version.rb
@@ -1,3 +1,0 @@
-module StashDatacite
-  VERSION = '0.0.5'.freeze
-end
